Refactor: Consolidate notification and notifications modules
