--- xconfess-backend/src/notification/notification.admin.controller.ts
@@ -1,67 +1,0 @@
-import {
-  Body,
-  Controller,
-  Get,
-  Param,
-  ParseIntPipe,
-  Post,
-  Query,
-  Req,
-  UseGuards,
-} from '@nestjs/common';
-import { JwtAuthGuard } from '../auth/jwt-auth.guard';
-import { AdminGuard } from '../auth/admin.guard';
-import { NotificationQueue } from './notification.queue';
-import { RateLimit } from '../auth/guard/rate-limit.decorator';
-import { RateLimitGuard } from '../auth/guard/rate-limit.guard';
-
-@Controller('admin/notifications')
-@UseGuards(JwtAuthGuard, AdminGuard)
-export class NotificationAdminController {
-  constructor(private readonly notificationQueue: NotificationQueue) {}
-
-  @Get('dlq')
-  async listDlqJobs(
-    @Query('page', new ParseIntPipe({ optional: true })) page?: number,
-    @Query('limit', new ParseIntPipe({ optional: true })) limit?: number,
-    @Query('failedAfter') failedAfter?: string,
-    @Query('failedBefore') failedBefore?: string,
-    @Query('search') search?: string,
-  ) {
-    return this.notificationQueue.listDlqJobs(page, limit, {
-      failedAfter,
-      failedBefore,
-      search,
-    });
-  }
-
-  @Post('dlq/:jobId/replay')
-  @UseGuards(RateLimitGuard)
-  @RateLimit(3, 60)
-  async replayDlqJob(
-    @Param('jobId') jobId: string,
-    @Body('reason') reason: string | undefined,
-    @Req() req: any,
-  ) {
-    const actorId = String(req.user?.id);
-    return this.notificationQueue.replayDlqJob(jobId, actorId, reason);
-  }
-
-  @Post('dlq/replay')
-  @UseGuards(RateLimitGuard)
-  @RateLimit(2, 60)
-  async replayDlqJobsBulk(
-    @Body()
-    body: {
-      limit?: number;
-      failedAfter?: string;
-      failedBefore?: string;
-      search?: string;
-      reason?: string;
-    },
-    @Req() req: any,
-  ) {
-    const actorId = String(req.user?.id);
-    return this.notificationQueue.replayDlqJobsBulk(actorId, body);
-  }
-}
